AMLS-4890 | Added/Removed spec files - Tests still need work
