+++ test/views/responsiblepeople/address/address/additional_extra_address_NonUKSpec.scala
@@ -14,20 +14,19 @@
  * limitations under the License.
  */
 
-package views.responsiblepeople
+package views.responsiblepeople.address.address
 
 import forms.{Form2, InvalidForm, ValidForm}
+import jto.validation.{Path, ValidationError}
+import models.autocomplete.NameValuePair
 import models.responsiblepeople.{PersonAddressUK, ResponsiblePersonAddress}
 import org.scalatest.MustMatchers
-import utils.AmlsSpec
-import jto.validation.Path
-import jto.validation.ValidationError
-import models.autocomplete.NameValuePair
 import play.api.i18n.Messages
+import utils.AmlsSpec
 import views.Fixture
 
 
-class additional_extra_addressSpec extends AmlsSpec with MustMatchers {
+class additional_extra_address_NonUKSpec extends AmlsSpec with MustMatchers {
 
   trait ViewFixture extends Fixture {
     implicit val requestWithToken = addToken(request)
@@ -41,7 +40,7 @@ class additional_extra_addressSpec extends AmlsSpec with MustMatchers {
 
     "have a back link" in new ViewFixture {
       val form2: ValidForm[ResponsiblePersonAddress] = Form2(ResponsiblePersonAddress(PersonAddressUK("","",None,None,""), None))
-      def view = views.html.responsiblepeople.additional_extra_address(form2, true, 1, None, "firstName lastName", countries)
+      def view = views.html.responsiblepeople.address.additional_extra_address_NonUK(form2, true, 1, None, "firstName lastName", countries)
       doc.getElementsByAttributeValue("class", "link-back") must not be empty
     }
 
@@ -49,69 +48,27 @@ class additional_extra_addressSpec extends AmlsSpec with MustMatchers {
 
       val form2: ValidForm[ResponsiblePersonAddress] = Form2(ResponsiblePersonAddress(PersonAddressUK("","",None,None,""), None))
 
-      def view = views.html.responsiblepeople.additional_extra_address(form2, true, 1, None, "firstName lastName", countries)
+      def view = views.html.responsiblepeople.address.additional_extra_address_NonUK(form2, true, 1, None, "firstName lastName", countries)
 
-      doc.title must startWith (Messages("responsiblepeople.additional_extra_address.title", "firstName lastName"))
+      doc.title must startWith (Messages("responsiblepeople.additional_extra_address_country.title", "firstName lastName"))
     }
 
     "have correct headings" in new ViewFixture {
 
       val form2: ValidForm[ResponsiblePersonAddress] = Form2(ResponsiblePersonAddress(PersonAddressUK("","",None,None,""), None))
 
-      def view = views.html.responsiblepeople.additional_extra_address(form2, true, 1, None, "firstName lastName", countries)
+      def view = views.html.responsiblepeople.address.additional_extra_address_NonUK(form2, true, 1, None, "firstName lastName", countries)
 
-      heading.html must be(Messages("responsiblepeople.additional_extra_address.heading", "firstName lastName"))
+      heading.html must be(Messages("responsiblepeople.additional_extra_address_country.heading", "firstName lastName"))
       subHeading.html must include(Messages("summary.responsiblepeople"))
 
     }
 
     "show errors in the correct locations" when {
-      "UK" in new ViewFixture {
-
-        val form2: InvalidForm = InvalidForm(Map.empty,
-          Seq(
-            (Path \ "isUK") -> Seq(ValidationError("not a message Key")),
-            (Path \ "addressLine1") -> Seq(ValidationError("second not a message Key")),
-            (Path \ "addressLine2") -> Seq(ValidationError("third not a message Key")),
-            (Path \ "addressLine3") -> Seq(ValidationError("fourth not a message Key")),
-            (Path \ "addressLine4") -> Seq(ValidationError("fifth not a message Key")),
-            (Path \ "postCode") -> Seq(ValidationError("sixth not a message Key"))
-          ))
-
-        def view = views.html.responsiblepeople.additional_extra_address(form2, true, 1, None, "firstName lastName", countries)
-
-        errorSummary.html() must include("not a message Key")
-        errorSummary.html() must include("second not a message Key")
-        errorSummary.html() must include("third not a message Key")
-        errorSummary.html() must include("fourth not a message Key")
-        errorSummary.html() must include("fifth not a message Key")
-        errorSummary.html() must include("sixth not a message Key")
-
-        doc.getElementById("isUK")
-          .getElementsByClass("error-notification").first().html() must include("not a message Key")
-
-        doc.getElementById("addressLine1").parent()
-          .getElementsByClass("error-notification").first().html() must include("second not a message Key")
-
-        doc.getElementById("addressLine2").parent()
-          .getElementsByClass("error-notification").first().html() must include("third not a message Key")
-
-        doc.getElementById("addressLine3").parent()
-          .getElementsByClass("error-notification").first().html() must include("fourth not a message Key")
-
-        doc.getElementById("addressLine4").parent()
-          .getElementsByClass("error-notification").first().html() must include("fifth not a message Key")
-
-        doc.getElementById("postCode").parent()
-          .getElementsByClass("error-notification").first().html() must include("sixth not a message Key")
-
-      }
-
       "non UK" in new ViewFixture {
 
         val form2: InvalidForm = InvalidForm(Map.empty,
           Seq(
-            (Path \ "isUK") -> Seq(ValidationError("not a message Key")),
             (Path \ "addressLineNonUK1") -> Seq(ValidationError("second not a message Key")),
             (Path \ "addressLineNonUK2") -> Seq(ValidationError("third not a message Key")),
             (Path \ "addressLineNonUK3") -> Seq(ValidationError("fourth not a message Key")),
@@ -119,17 +76,12 @@ class additional_extra_addressSpec extends AmlsSpec with MustMatchers {
             (Path \ "country") -> Seq(ValidationError("sixth not a message Key"))
           ))
 
-        def view = views.html.responsiblepeople.additional_extra_address(form2, true, 1, None, "firstName lastName", countries)
-
-        errorSummary.html() must include("not a message Key")
+        def view = views.html.responsiblepeople.address.additional_extra_address_NonUK(form2, true, 1, None, "firstName lastName", countries)
         errorSummary.html() must include("second not a message Key")
         errorSummary.html() must include("third not a message Key")
         errorSummary.html() must include("fourth not a message Key")
         errorSummary.html() must include("fifth not a message Key")
         errorSummary.html() must include("sixth not a message Key")
-
-        doc.getElementById("isUK")
-          .getElementsByClass("error-notification").first().html() must include("not a message Key")
 
         doc.getElementById("addressLineNonUK1").parent()
           .getElementsByClass("error-notification").first().html() must include("second not a message Key")
